+++ CHANGELOG.md
@@ -19,6 +19,7 @@ develop
 
 - Fix an issue where documents were marked as “Edited” just after opening file if “link URL” option is enabled.
 - Fix an issue where URL link was not applied to pasted text.
+- Fix an issue where find-all highlight wasn't removed if find panel is closed before closing find result view.